requirements.txt
streamlit
matplotlib
seaborn
urlextract
wordcloud
pandas
emoji
psycopg2-binary
textblob
pyperclip